refactor: extract SpawnedCommand to separate file
