+++ spec/support/command_helpers.rb
@@ -2,42 +2,8 @@
 
 require "securerandom"
 require "pty"
-require "expect"
 
-class SpawnedCommand
-  attr_reader :output, :input, :pid
-
-  DEFAULT_TIMEOUT = 120
-
-  def initialize(output, input, pid)
-    @output = output
-    @input = input
-    @pid = pid
-  end
-
-  def wait_for(regex, timeout: DEFAULT_TIMEOUT)
-    result = nil
-    output.expect(regex, timeout) do |matches|
-      result = matches&.first
-    end
-
-    raise "Timed out waiting for #{regex.inspect} after #{timeout} seconds" if result.nil?
-
-    result
-  end
-
-  def wait_for_prompt
-    wait_for(/[$#>]/)
-  end
-
-  def type(string)
-    input.puts("#{string}\n")
-  end
-
-  def kill
-    Process.kill("INT", pid)
-  end
-end
+require_relative "spawned_command"
 
 module CommandHelpers # rubocop:disable Metrics/ModuleLength
   module_function
